agri_iot/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
asgiref==3.9.1
    # via
    #   django
    #   django-cors-headers
boto3==1.40.1
    # via -r requirements.in
botocore==1.40.1
    # via
    #   boto3
    #   s3transfer
cffi==1.17.1
    # via cryptography
cryptography==43.0.3
    # via -r requirements.in
dj-database-url==2.1.0
    # via -r requirements.in
django==4.2.23
    # via
    #   -r requirements.in
    #   dj-database-url
    #   django-cors-headers
    #   django-debug-toolbar
    #   django-redis
    #   django-storages
django-cors-headers==4.3.1
    # via -r requirements.in
django-debug-toolbar==6.0.0
    # via -r requirements.in
django-redis==5.4.0
    # via -r requirements.in
django-storages==1.14.2
    # via -r requirements.in
gunicorn==23.0.0
    # via -r requirements.in
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
packaging==25.0
    # via gunicorn
pdf2image==1.17.0
    # via -r requirements.in
pillow==11.3.0
    # via
    #   -r requirements.in
    #   pdf2image
psycopg2-binary==2.9.9
    # via -r requirements.in
pycparser==2.22
    # via cffi
pyjwt==2.10.1
    # via -r requirements.in
python-dateutil==2.9.0
    # via
    #   -r requirements.in
    #   botocore
python-json-logger==3.3.0
    # via -r requirements.in
pyyaml==6.0.1
    # via -r requirements.in
redis==5.0.1
    # via
    #   -r requirements.in
    #   django-redis
s3transfer==0.13.1
    # via boto3
six==1.17.0
    # via python-dateutil
sqlparse==0.5.3
    # via
    #   django
    #   django-debug-toolbar
typing-extensions==4.14.1
    # via dj-database-url
urllib3==2.5.0
    # via botocore
whitenoise==6.6.0
    # via -r requirements.in
